requirements.txt
backcall==0.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.0
Django==2.1.14
django-filter==2.2.0
djangorestframework==3.10.3
dnspython==1.16.0
dnspython3==1.15.0
Flask==1.1.1
idna==2.8
ipython==7.7.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
itypes==1.1.0
jedi==0.14.1
Jinja2==2.10.1
MarkupSafe==1.1.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prettytable==0.7.2
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.25.3
uWSGI==2.0.18
wcwidth==0.1.7
Werkzeug==0.15.5
